requirements.txt
#Using Python 3.10.13, 3.12.4(en miniconda)
pandas==2.2.2
#scipy==1.14.0
streamlit==1.36.0
plotly-express
nbformat==5.10.4